polymer-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymer-mcp"
version = "0.1.0"
description = "Master Coder Protocol — Polymers: data, physics, simulation, ML, and CLI."
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "MCP", email = "example@example.com" }]
keywords = ["polymers", "materials", "MCP", "BigSMILES", "scientific-method", "CLI", "Hydra"]
license = { text = "MIT" }

dependencies = [
  "typer[all]>=0.12.3",
  "pydantic>=2.6",
  "jsonschema>=4.22",
  "PyYAML>=6.0.2",
  "pandas>=2.2",
  "numpy>=1.26",
  "scikit-learn>=1.5",
  "joblib>=1.4",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "matplotlib>=3.8",
  "jinja2>=3.1",
  "mlflow>=2.14",
]

[project.optional-dependencies]
all = [
  "polymer-mcp",
]

[project.scripts]
polymerlab = "polymerlab.cli.app:app"

[tool.setuptools.packages.find]
where = ["src"]
include = ["polymerlab*"]
